refactor testcontainers to use index.ts structure
